utah-forage-map/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
authlib
passlib
bcrypt
httpx
crawl4ai
schedule
